requirements.txt
Flask-WTF
email_validator
flask-admin
flask-login